Refactoring of gui items for ease of extensibility and edits
