+++ gui/doi/__init__.py
@@ -1,0 +1,3 @@
+from gui.doi.page import DoiPage
+
+__all__ = ["DoiPage"]
