+++ gui/graph/__init__.py
@@ -1,0 +1,4 @@
+from gui.graph.page import GraphPage
+from gui.graph.view import GraphView
+
+__all__ = ["GraphPage", "GraphView"]
