+++ gui/home/__init__.py
@@ -1,0 +1,3 @@
+from gui.home.page import HomePage
+
+__all__ = ["HomePage"]
