+++ gui/library/__init__.py
@@ -1,0 +1,3 @@
+from gui.library.page import LibraryPage
+
+__all__ = ["LibraryPage"]
