+++ gui/projects/__init__.py
@@ -1,0 +1,3 @@
+from gui.projects.page import ProjectsPage, NoteEditorDialog
+
+__all__ = ["ProjectsPage", "NoteEditorDialog"]
