+++ gui/setup/__init__.py
@@ -1,0 +1,3 @@
+from gui.setup.page import SetupPage
+
+__all__ = ["SetupPage"]
