+++ gui/views/__init__.py
@@ -1,0 +1,5 @@
+from gui.views.markdown_view import MarkdownView
+from gui.views.tex_view import TexView
+from gui.views.pdf_window import PdfWindow
+
+__all__ = ["MarkdownView", "TexView", "PdfWindow"]
