+++ gui/views/markdown_view.py
