+++ gui/views/tex_view.py
