+++ gui/views/web/fonts/KaTeX_Caligraphic-Regular.woff2
