+++ gui/views/web/fonts/KaTeX_Fraktur-Bold.woff2
