+++ gui/views/web/fonts/KaTeX_Fraktur-Regular.woff2
